==> submissions/WUSTL_HPRC_HiFi_Year2_TopUp/1-wrangle-hifi_qc.ipynb <==
import pandas as pd
import glob

wustl_hifi_y2_topup = pd.read_csv('WUSTL_HiFi_Year2_TopUp.csv',header=None)

wustl_hifi_y2_topup = wustl_hifi_y2_topup[wustl_hifi_y2_topup[0].str.contains('hifi_reads')]

wustl_hifi_y2_topup['sample_id'] = [sample.split('/')[-3] for sample in wustl_hifi_y2_topup[0].tolist()]

wustl_hifi_y2_topup.shape

other_reads = []
for sample in [sample.split('/')[-3] for sample in wustl_hifi_y2_topup[0].tolist()]:
	other_reads_path = glob.glob('/private/groups/hprc/human-pangenomics/1000G_local/child/cram_to_bam/'+sample+'/*/*/*/*')
	if other_reads_path == []:
		other_reads_path = glob.glob('/private/groups/hprc/1000G_local/child/fastq/' + sample + '.*')
		other_reads.append([other_reads_path[0]])
	else:
		other_reads.append([other_reads_path[0]])
wustl_hifi_y2_topup['other_reads'] = other_reads
wustl_hifi_y2_topup['perform_methylation_check'] = 'true'


wustl_hifi_y2_topup.rename(columns={0:'hifi_reads'},inplace=True)

wustl_hifi_y2_topup['sample_id'] = ['-'.join(item) for item in zip(wustl_hifi_y2_topup['sample_id'].tolist(), [sample.split('/')[-1].split('.')[0] for sample in wustl_hifi_y2_topup['hifi_reads'].tolist()])]

wustl_hifi_y2_topup[['sample_id','hifi_reads', 'other_reads', 'perform_methylation_check']].to_csv('hifi_qc/hifi_qc_sample_files.csv', index=False)
