==> submissions/WUSTL_HPRC_HiFi_Year2_TopUp/4-evaluate-hifi_qc.ipynb <==
import pandas as pd

hifi_df = pd.read_csv('hifi_qc/hifi_qc_samples_output_table.csv')
hifi_qc = pd.concat([pd.read_csv(qc,sep='\t') for qc in hifi_df['hifi_qc_summary'].tolist()])
hifi_qc['sample_id'] = [sample.split('-')[0] for sample in hifi_qc['filename'].tolist()]
# hifi_qc['coverage'] = [bp/3.1 for bp in hifi_qc['total_Gbp'].tolist()]

hifi_qc.drop(columns=['all_kinetics_flag', 'keep_kinetics_flag', 'hifi_kinetics_tag', 'PP_PRIMROSE', 'fi_tag', 'ri_tag','fp_tag', 'rp_tag'],inplace=True)

hifi_qc

hifi_qc.groupby('sample_id')['total_Gbp'].sum().div(3.1).reset_index(name='coverage')

hifi_qc.to_csv('WUSTL_HPRC_HiFi_Year2_TopUp_sample_file_hifi_qc.csv',index=False)
