==> submissions/WUSTL_HPRC_HiFi_Year2_TopUp/5-release.ipynb <==
import pandas as pd

hifi_output_df = pd.read_csv('/private/groups/hprc/human-pangenomics/HPRC_metadata/submissions/WUSTL_HPRC_HiFi_Year2_TopUp/hifi_qc/hifi_qc_samples_output_table.csv') # submission file


def aws_working(file_paths):
	new_path_list = []
	for data in file_paths:
		sample_id = data.split('/')[-3]
		new_path = f"s3://human-pangenomics/working/HPRC/{sample_id}/raw_data/PacBio_HiFi/{data.split('/')[-1]}"
		new_path_list.append(new_path)
	return new_path_list
aws_working = aws_working(hifi_output_df['hifi_reads'].tolist())

aws_df = pd.DataFrame(hifi_output_df['hifi_reads'])
aws_df['release'] = aws_working

aws_df.to_csv('WUSTL_HiFi_Year2_TopUp.transfer.csv', header=None, index=None)
